chore: added filtersSection component for the filters
