+++ pkg/app/src/pages/app/organisations/index.tsx
@@ -4,18 +4,8 @@ import Typography from '@mui/material/Typography'
 import { ItemList } from 'components/OrganisationCard/itemList'
 import { Layout } from 'src/layouts/default/layout'
 import { Body, useBodiesQuery } from '@gamedao-haiku/graphql/dist'
-import {
-	Button,
-	Container,
-	createSvgIcon,
-	FormControl,
-	Grid,
-	InputAdornment,
-	NativeSelect,
-	TextField,
-} from '@mui/material'
-import { makeStyles } from '@mui/styles'
-import FilterListIcon from '@mui/icons-material/FilterList'
+import { Button, Container, createSvgIcon, Grid } from '@mui/material'
+import FiltersSection from 'components/OrganisationCard/modules/filtersSection'
 
 const PlusIcon = createSvgIcon(
 	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
@@ -28,40 +18,7 @@ const PlusIcon = createSvgIcon(
 	'Plus',
 )
 
-const SearchIcon = createSvgIcon(
-	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
-		={' '}
-		<path
-			fillRule="evenodd"
-			d="M8 4a4 4 0 100 8 4 4 0 000-8zM2 8a6 6 0 1110.89 3.476l4.817 4.817a1 1 0 01-1.414 1.414l-4.816-4.816A6 6 0 012 8z"
-			clipRule="evenodd"
-		/>
-	</svg>,
-	'Search',
-)
-
 export function OrganisationPage() {
-	const useStyles = makeStyles({
-		field: {
-			'& fieldset': {
-				borderRadius: 8,
-			},
-		},
-		formControl: {
-			'& .MuiInput-underline:hover:not(.Mui-disabled):before': {
-				borderColor: 'blue',
-			},
-		},
-		select: {
-			'&&&:before': {
-				borderBottom: 'none',
-			},
-			'&&:after': {
-				borderBottom: 'none',
-			},
-		},
-	})
-	const classes = useStyles()
 	const { loading, error, data } = useBodiesQuery()
 	useEffect(() => {
 		if (error) {
@@ -94,90 +51,9 @@ export function OrganisationPage() {
 								</Button>
 							</Grid>
 						</Grid>
-						<Box
-							sx={{
-								m: -1,
-								mt: 3,
-							}}
-						>
-							<Box
-								sx={{
-									flexGrow: 1,
-									m: 1.5,
-								}}
-							>
-								<Grid container spacing={3}>
-									<Grid item xs={12} md={4}>
-										<TextField
-											id="outlined-basic"
-											size="small"
-											className={classes.field}
-											fullWidth
-											InputProps={{
-												startAdornment: (
-													<InputAdornment position="start">
-														<SearchIcon fontSize="small" />
-													</InputAdornment>
-												),
-											}}
-											placeholder="Search Organisations…"
-										/>
-									</Grid>
-									<Grid item xs={0} md={4}></Grid>
-									<Grid item xs={12} md={4}>
-										<Box
-											sx={{
-												display: 'flex',
-												alignItems: 'center',
-											}}
-										>
-											<Box
-												sx={{
-													display: 'flex',
-													justifyContent: 'space-around',
-													alignItems: 'center',
-												}}
-											>
-												<Typography sx={{ fontWeight: '700' }} variant={'body2'}>
-													Filters
-												</Typography>
-												<FilterListIcon fontSize={'small'} sx={{ ml: 1 }} />
-											</Box>
-											<Box
-												sx={{
-													flexGrow: 1,
-													display: 'flex',
-													justifyContent: 'space-evenly',
-													alignItems: 'center',
-												}}
-											>
-												<Typography sx={{ fontWeight: '700' }} variant={'body2'}>
-													Sort By:
-												</Typography>
-												<Box sx={{ minWidth: 120 }}>
-													<FormControl fullWidth className={classes.formControl}>
-														<NativeSelect
-															className={classes.select}
-															defaultValue={30}
-															inputProps={{
-																name: 'sort',
-																id: 'uncontrolled-native',
-																style: { fontSize: 14 },
-															}}
-														>
-															<option value={10}>Member: High-Low</option>
-															<option value={20}>Member: Low-High</option>
-														</NativeSelect>
-													</FormControl>
-												</Box>
-											</Box>
-										</Box>
-									</Grid>
-								</Grid>
-							</Box>
-						</Box>
+						<FiltersSection />
 					</Box>
-					<ItemList items={data?.bodies?.slice() as Body[]} loading={loading}></ItemList>
+					<ItemList items={data?.bodies?.slice() as Body[]} loading={loading} />
 				</Container>
 			</Box>
 		</Layout>
